Rename, use example_data for file system tests
