+++ libs/langchain-oracle/src/index.ts
@@ -1,5 +1,5 @@
-export * from "./doc_loader.js";
+export * from "./document_loaders.js";
 export * from "./embeddings.js";
 export * from "./summary.js";
-export * from "./vectorstore.js";
 export * from "./text_splitter.js";
+export * from "./vectorstores.js";
